requirements.txt
python-rtmidi==1.4.9
PyYAML==6.0
tk==0.1.0
